+++ tests/mockers.py
@@ -103,6 +103,10 @@ class GitRepositorySandbox:
         self.execute(f'git remote add {remote} {url}')
         return self
 
+    def remove_remote(self, remote: str) -> "GitRepositorySandbox":
+        self.execute(f'git remote remove {remote}')
+        return self
+
     def add_git_config_key(self, key: str, value: str) -> "GitRepositorySandbox":
         self.execute(f'git config {key} {value}')
         return self